Backend/django-pgvector/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.30",
    "django>=5.2.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "onnxruntime>=1.22.0",
    "pgvector>=0.4.1",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
]
